+++ src/collective/documentgenerator/helper/base.py
@@ -106,6 +106,14 @@ class DocumentGenerationHelperView(object):
             state = wtool.getTitleForStateOnType(state, obj.portal_type)
         return state
 
+    def context_var(self, name, default=''):
+        """ Test if a context variable is defined and return it or return default """
+        ctx = self.appy_renderer.contentParser.env.context
+        if name in ctx:
+            return ctx[name]
+        else:
+            return default
+
 
 class DisplayProxyObject(object):
     """